Runner-Up/requirements.txt
google-generativeai
rich
SpeechRecognition
pyttsx3
sounddevice
numpy
scipy
